# Conflicts:
#	src/app/pages/common/entity/utils.ts

+++ src/app/pages/applications/chart-releases/chart-releases.component.ts
@@ -8,7 +8,6 @@ import { DialogService, SystemGeneralService, WebSocketService } from '../../../
 import { ApplicationsService } from '../applications.service';
 import { ModalService } from '../../../services/modal.service';
 import { EntityJobComponent } from '../../common/entity/entity-job/entity-job.component';
-import { EntityUtils } from '../../common/entity/utils';
 import { DialogFormConfiguration } from '../../common/entity/entity-dialog/dialog-form-configuration.interface';
 import { ChartReleaseEditComponent } from '../forms/chart-release-edit.component';
 import { CommonUtils } from 'app/core/classes/common-utils';
@@ -16,10 +15,9 @@ import { ChartFormComponent } from '../forms/chart-form.component';
 import { EmptyConfig, EmptyType } from '../../common/entity/entity-empty/entity-empty.component';
 
 import  helptext  from '../../../helptext/apps/apps';
-import { EntityToolbarComponent } from 'app/pages/common/entity/entity-toolbar/entity-toolbar.component';
 import { CoreService, CoreEvent } from 'app/core/services/core.service';
-import { BulkOptionsComponent } from '../forms/bulk-options.component';
 import { Router } from '@angular/router';
+import { ChartEventsDialog } from '../dialogs/chart-events/chart-events-dialog.component';
 
 @Component({
   selector: 'app-charts',
@@ -136,11 +134,11 @@ export class ChartReleasesComponent implements OnInit {
     if (evt.data.event_control == 'filter') {
       this.filterString = evt.data.filter;
       this.filerChartItems();
+
     } else if (evt.data.event_control == 'bulk') {
-      this.bulkOptions();
+      this.onBulkAction(evt.data.bulk.value);
     }
-  }
-  
+  }  
 
   viewCatalog() {
     this.updateTab.emit({name: 'SwitchTab', value: '0'});
@@ -230,11 +228,15 @@ export class ChartReleasesComponent implements OnInit {
                 let chartObj = {
                   name: chart.name,
                   catalog: chart.catalog,
+                  catalog_train: chart.catalog_train,
                   status: chart.status,
                   version: chart.chart_metadata.version,
+                  human_version: chart.human_version,
+                  human_latest_version: chart.human_latest_version,
+                  container_images_update_available: chart.container_images_update_available,
                   latest_version: chart.chart_metadata.latest_chart_version,
                   description: chart.chart_metadata.description,
-                  update: chart.update_available,
+                  update_available: chart.update_available,
                   chart_name: chart.chart_metadata.name,
                   repository: chart.config.image.repository,
                   tag: chart.config.image.tag,
@@ -244,6 +246,7 @@ export class ChartReleasesComponent implements OnInit {
                   count: `${chart.pod_status.available}/${chart.pod_status.desired}`,
                   desired: chart.pod_status.desired,
                   history: !(_.isEmpty(chart.history)),
+                  selected: false,
                 };
         
                 let ports = [];
@@ -349,33 +352,59 @@ export class ChartReleasesComponent implements OnInit {
     }
   }
 
-  onBulkAction(checkedItems: any[], actionName: string) {
-    if (actionName === 'delete') {
-      this.bulkDelete(checkedItems);      
-    } else {
-      checkedItems.forEach(name => {
-        switch (actionName) {
-          case 'start':
-            this.start(name);
-            break;
-          case 'stop':
-            this.stop(name);
-            break;
-        }
-      });
+  getSelectedItems() {
+    const selectedItems = [];
+    this.filteredChartItems.forEach(element => {
+      if (element.selected) {
+        selectedItems.push(element.name);
+      }
+    });   
+    return selectedItems;
+  }
   
-      this.translate.get(helptext.bulkActions.finished).subscribe(msg => {
-        this.dialogService.Info(helptext.choosePool.success, msg,
-          '500px', 'info', true);
-      })
-    }    
+  checkAll(checkedItems) {
+    let selectAll = true;
+    if (checkedItems.length == this.filteredChartItems.length) {
+      selectAll = false;
+    }
+
+    this.filteredChartItems.forEach(item => {
+      item.selected = selectAll;
+    });
   }
 
-  bulkOptions() {
-    const bulkOptionsForm = new BulkOptionsComponent(this.modalService, this.appService);
-    bulkOptionsForm.setParent(this);
-
-    this.modalService.open('slide-in-form', bulkOptionsForm, "Bulk Options");
+  onBulkAction(actionName: string) {
+    const checkedItems = this.getSelectedItems();
+    
+    if (actionName === 'select_all') {
+      this.checkAll(checkedItems);
+    } else {
+      if (checkedItems.length > 0) {
+        if (actionName === 'delete') {
+          this.bulkDelete(checkedItems);      
+        } else {
+          checkedItems.forEach(name => {
+            switch (actionName) {
+              case 'start':
+                this.start(name);
+                break;
+              case 'stop':
+                this.stop(name);
+                break;
+            }
+          });
+    
+          this.translate.get(helptext.bulkActions.finished).subscribe(msg => {
+            this.dialogService.Info(helptext.bulkActions.success, msg,
+              '500px', 'info', true);
+          });
+        }
+      } else {
+        this.translate.get(helptext.bulkActions.no_selected).subscribe(msg => {
+          this.dialogService.errorReport(helptext.bulkActions.error, msg);
+        });
+      }
+    } 
   }
 
   delete(name: string) {
@@ -508,5 +537,17 @@ export class ChartReleasesComponent implements OnInit {
       });
       entityDialog.formGroup.controls['containers'].setValue(containers[0]);
     })
+  }
+
+  showChartEvents(name: string) {
+    const catalogApp = this.chartItems[name];
+    if (catalogApp) {
+      let dialogRef = this.mdDialog.open(ChartEventsDialog, {
+        width: '686px',
+        maxWidth: '686px',
+        data: catalogApp,
+        disableClose: false,
+      });
+    }
   }
 }
